Update functionality complete, big cleanup and refactor
