+++ src/components/organisms/index.js
@@ -1,3 +1,5 @@
 export { default as Navbar } from './Navbar'
-export { default as NewLoreModal } from './NewLoreModal'
+export { default as LoreFormModal } from './LoreFormModal'
 export { default as LoreMenu } from './LoreMenu'
+export { default as DeleteLoreMenuItem } from './DeleteLoreMenuItem'
+export { default as UpdateLoreMenuItem } from './UpdateLoreMenuItem'
