+++ src/utils/constants.js
@@ -1,0 +1,4 @@
+/**
+ * Base form fields
+ */
+export const loreFieldsArr = ['title', 'subtitle', 'game', 'text']
